[Refactor] Refactor Packaging
