+++ app/src/main/java/org/sopt/sample/data/repository/ReqresRepositoryImpl.kt
@@ -1,7 +1,8 @@
-package org.sopt.sample.data.model.repository
+package org.sopt.sample.data.repository
 
 import org.sopt.sample.application.ApiFactory
 import org.sopt.sample.data.model.dto.ResponseReqresListDTO
+import org.sopt.sample.domain.repository.ReqresRepository
 import retrofit2.Response
 
 class ReqresRepositoryImpl : ReqresRepository {
