+++ app/src/main/java/org/sopt/sample/presentation/home/HomeActivity.kt
@@ -4,7 +4,7 @@ import android.os.Bundle
 import androidx.fragment.app.Fragment
 import androidx.lifecycle.ViewModelProvider
 import org.sopt.sample.R
-import org.sopt.sample.data.model.repository.ReqresRepositoryImpl
+import org.sopt.sample.data.repository.ReqresRepositoryImpl
 import org.sopt.sample.databinding.ActivityHomeBinding
 import org.sopt.sample.presentation.base.BindingActivity
 import org.sopt.sample.presentation.home.fragment.GalleryFragment
