+++ app/src/main/java/org/sopt/sample/presentation/home/viewmodel/ReqresListViewModelFactory.kt
@@ -2,7 +2,7 @@ package org.sopt.sample.presentation.home.viewmodel
 
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.ViewModelProvider
-import org.sopt.sample.data.model.repository.ReqresRepositoryImpl
+import org.sopt.sample.data.repository.ReqresRepositoryImpl
 
 @Suppress("UNCHECKED_CAST")
 class ReqresListViewModelFactory( // 의존성 주입은 추후에 더 공부해서 하는걸로ㅠㅠ
